+++ tests/app/src/js/ReactFormSelect.js
@@ -6,16 +6,13 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @author Beñat Espiña <benatespina@gmail.com>
  * @author Mikel Tuesta <mikeltuesta@gmail.com>
  */
 
-import loadScript from './loadScript';
-import waitImagesLoadInDomNode from './waitImagesLoadInDomNode';
-import getHtmlLang from './getHtmlLang';
+import {onDomReady} from 'lin3s-event-bus';
 
-export {
-  loadScript,
-  waitImagesLoadInDomNode,
-  getHtmlLang
-}
+const onReady = () => {
+
+};
+
+onDomReady(onReady);